backend/requirements.txt
flask
flask-cors
qdrant-client
sentence-transformers
pandas
requests
groq
dotenv
tavily-python
